+++ package.json
@@ -28,11 +28,11 @@
     "faker": "4.1.0"
   },
   "scripts": {
-    "start": "node server.js",
+    "start": "node server",
     "postinstall": "npm run webpack",
     "webpack": "webpack",
     "webpack:dev": "npm run webpack -- --mode=development --watch",
-    "start:dev": "npm run webpack:dev & DATABASE_URL=postgres://localhost/grace-shopper nodemon server.js --ignore src/ --ignore dist/"
+    "start:dev": "npm run webpack:dev & DATABASE_URL=postgres://localhost/grace-shopper nodemon server --ignore src/ --ignore dist/"
   },
   "keywords": [],
   "author": "",
